Converted CSChurches into a class
